Reorginized tests to ci-build and nightly-build folders
