+++ test/ci-build/auth/anonymous-user.e2e-spec.ts
@@ -1,7 +1,6 @@
-import { graphqlRequest, graphqlRequestAuth } from '../utils/graphql.request';
-import { TestUser } from '../utils/token.helper';
-import '../utils/array.matcher';
-import { appSingleton } from '../utils/app.singleton';
+import { graphqlRequest, graphqlRequestAuth } from '../../utils/graphql.request';
+import '../../utils/array.matcher';
+import { appSingleton } from '../../utils/app.singleton';
 import {
   name,
   hostMembers,
@@ -81,7 +80,7 @@ import {
   templatesName,
   templatesDescription,
   templatesUsersName,
-} from '../utils/queries';
+} from './queries';
 
 import {
   createOrganisationMutation,
@@ -118,7 +117,7 @@ import {
   createReferenceOnContextVariables,
   createTagsetOnProfileMutation,
   createTagsetOnProfileVariables,
-} from '../utils/create-mutations';
+} from './create-mutations';
 
 import {
   updateUserMutation,
@@ -153,7 +152,7 @@ import {
   addChallengeLeadToOrganisationVariables,
   removeUserFromGroupMutation,
   removeUserFromGroupVariables,
-} from '../utils/update-mutations';
+} from './update-mutations';
 
 import {
   removeUserMutation,
@@ -166,7 +165,7 @@ import {
   removeActorVariables,
   removeActorGroupMutation,
   removeActorGroupVariables,
-} from '../utils/remove-mutations';
+} from './remove-mutations';
 
 const notAuthorizedCode = `"code":"UNAUTHENTICATED"`;
 
@@ -178,13 +177,82 @@ afterAll(async () => {
   if (appSingleton.Instance.app) await appSingleton.Instance.teardownServer();
 });
 
-describe('DDT community admin user - queries - authorized', () => {
+describe('DDT anonymous user - queries - Not authorized', () => {
+  // Arrange
+  test.each`
+    query                               | expected
+    ${hostGroups}                       | ${notAuthorizedCode}
+    ${hostMembers}                      | ${notAuthorizedCode}
+    ${usersName}                        | ${notAuthorizedCode}
+    ${usersAccountUPN}                  | ${notAuthorizedCode}
+    ${usersProfile}                     | ${notAuthorizedCode}
+    ${usersMemberofGroupsName}          | ${notAuthorizedCode}
+    ${usersMemberofChallengesName}      | ${notAuthorizedCode}
+    ${usersMemberofOrganisationsName}   | ${notAuthorizedCode}
+    ${userName}                         | ${notAuthorizedCode}
+    ${userAccountUPN}                   | ${notAuthorizedCode}
+    ${userProfile}                      | ${notAuthorizedCode}
+    ${userMemberofGroupsName}           | ${notAuthorizedCode}
+    ${userMemberofChallengesName}       | ${notAuthorizedCode}
+    ${userMemberofOrganisationsName}    | ${notAuthorizedCode}
+    ${usersById}                        | ${notAuthorizedCode}
+    ${groupsName}                       | ${notAuthorizedCode}
+    ${groupsFocalPointName}             | ${notAuthorizedCode}
+    ${groupsProfile}                    | ${notAuthorizedCode}
+    ${groupsMembersName}                | ${notAuthorizedCode}
+    ${groupsParentChallenge}            | ${notAuthorizedCode}
+    ${groupsParentEcoverse}             | ${notAuthorizedCode}
+    ${groupsParentOpportunity}          | ${notAuthorizedCode}
+    ${groupsWithTagName}                | ${notAuthorizedCode}
+    ${groupsWithTagFocalPointName}      | ${notAuthorizedCode}
+    ${groupsWithTagProfile}             | ${notAuthorizedCode}
+    ${groupsWithTagMembersName}         | ${notAuthorizedCode}
+    ${groupsWithTagParentChallenge}     | ${notAuthorizedCode}
+    ${groupsWithTagParentEcoverse}      | ${notAuthorizedCode}
+    ${groupsWithTagParentOpportunity}   | ${notAuthorizedCode}
+    ${challengesLeadOrganisationGroups} | ${notAuthorizedCode}
+    ${challengesContributors}           | ${notAuthorizedCode}
+    ${challengesGroups}                 | ${notAuthorizedCode}
+    ${challengeLeadOrganisationGroups}  | ${notAuthorizedCode}
+    ${challengeContributors}            | ${notAuthorizedCode}
+    ${challengeGroups}                  | ${notAuthorizedCode}
+    ${opportunitiesContributors}        | ${notAuthorizedCode}
+    ${opportunitiesGroups}              | ${notAuthorizedCode}
+    ${projectsName}                     | ${notAuthorizedCode}
+    ${projectsTextId}                   | ${notAuthorizedCode}
+    ${projectsDescription}              | ${notAuthorizedCode}
+    ${projectsState}                    | ${notAuthorizedCode}
+    ${projectsTagset}                   | ${notAuthorizedCode}
+    ${projectsAspects}                  | ${notAuthorizedCode}
+  `(
+    "should expect: '$expected' for query: '$query'",
+    async ({ query, expected }) => {
+      // Act
+      const requestParamsQueryData = {
+        operationName: null,
+        query: `${query}`,
+        variables: null,
+      };
+      const response = await graphqlRequest(requestParamsQueryData);
+      let responseData = JSON.stringify(response.body).replace('\\', '');
+      // 
+
+      // Assert
+      expect(response.status).toBe(200);
+      expect(responseData).toContain(expected);
+    }
+  );
+});
+
+// Failing scenarios BUG: https://app.zenhub.com/workspaces/cherrytwist-5ecb98b262ebd9f4aec4194c/issues/cherrytwist/server/578
+// ${opportunitiesProjectsId}          | ${notAuthorizedCode}
+// ${opportunitiesProjectsAspectsId}   | ${notAuthorizedCode}
+
+describe('DDT anonymous user - queries - authorized', () => {
   // Arrange
   test.each`
     query                                 | expected
     ${name}                               | ${notAuthorizedCode}
-    ${hostGroups}                         | ${notAuthorizedCode}
-    ${hostMembers}                        | ${notAuthorizedCode}
     ${hostProfile}                        | ${notAuthorizedCode}
     ${contextTagline}                     | ${notAuthorizedCode}
     ${contextBackground}                  | ${notAuthorizedCode}
@@ -192,69 +260,28 @@ describe('DDT community admin user - queries - authorized', () => {
     ${contextWho}                         | ${notAuthorizedCode}
     ${contextImpact}                      | ${notAuthorizedCode}
     ${contextReferencesName}              | ${notAuthorizedCode}
-    ${usersName}                          | ${notAuthorizedCode}
-    ${usersAccountUPN}                    | ${notAuthorizedCode}
-    ${usersProfile}                       | ${notAuthorizedCode}
-    ${usersMemberofGroupsName}            | ${notAuthorizedCode}
-    ${usersMemberofChallengesName}        | ${notAuthorizedCode}
-    ${usersMemberofOrganisationsName}     | ${notAuthorizedCode}
-    ${userName}                           | ${notAuthorizedCode}
-    ${userAccountUPN}                     | ${notAuthorizedCode}
-    ${userProfile}                        | ${notAuthorizedCode}
-    ${userMemberofGroupsName}             | ${notAuthorizedCode}
-    ${userMemberofChallengesName}         | ${notAuthorizedCode}
-    ${userMemberofOrganisationsName}      | ${notAuthorizedCode}
-    ${usersById}                          | ${notAuthorizedCode}
-    ${groupsName}                         | ${notAuthorizedCode}
-    ${groupsFocalPointName}               | ${notAuthorizedCode}
-    ${groupsProfile}                      | ${notAuthorizedCode}
-    ${groupsMembersName}                  | ${notAuthorizedCode}
-    ${groupsParentChallenge}              | ${notAuthorizedCode}
-    ${groupsParentEcoverse}               | ${notAuthorizedCode}
-    ${groupsParentOpportunity}            | ${notAuthorizedCode}
-    ${groupsWithTagName}                  | ${notAuthorizedCode}
-    ${groupsWithTagFocalPointName}        | ${notAuthorizedCode}
-    ${groupsWithTagProfile}               | ${notAuthorizedCode}
-    ${groupsWithTagMembersName}           | ${notAuthorizedCode}
-    ${groupsWithTagParentChallenge}       | ${notAuthorizedCode}
-    ${groupsWithTagParentEcoverse}        | ${notAuthorizedCode}
-    ${groupsWithTagParentOpportunity}     | ${notAuthorizedCode}
     ${challengesName}                     | ${notAuthorizedCode}
     ${challengesTextId}                   | ${notAuthorizedCode}
     ${challengesState}                    | ${notAuthorizedCode}
     ${challengesContext}                  | ${notAuthorizedCode}
     ${challengesLeadOrganisation}         | ${notAuthorizedCode}
-    ${challengesLeadOrganisationGroups}   | ${notAuthorizedCode}
-    ${challengesContributors}             | ${notAuthorizedCode}
     ${challengesTagsets}                  | ${notAuthorizedCode}
-    ${challengesGroups}                   | ${notAuthorizedCode}
     ${challengesOpportunities}            | ${notAuthorizedCode}
     ${challengeName}                      | ${notAuthorizedCode}
     ${challengeTextId}                    | ${notAuthorizedCode}
     ${challengeState}                     | ${notAuthorizedCode}
     ${challengeContext}                   | ${notAuthorizedCode}
     ${challengeLeadOrganisation}          | ${notAuthorizedCode}
-    ${challengeLeadOrganisationGroups}    | ${notAuthorizedCode}
-    ${challengeContributors}              | ${notAuthorizedCode}
     ${challengeTagsets}                   | ${notAuthorizedCode}
-    ${challengeGroups}                    | ${notAuthorizedCode}
     ${challengeOpportunities}             | ${notAuthorizedCode}
     ${opportunitiesName}                  | ${notAuthorizedCode}
     ${opportunitiesTextId}                | ${notAuthorizedCode}
     ${opportunitiesState}                 | ${notAuthorizedCode}
     ${opportunitiesContext}               | ${notAuthorizedCode}
-    ${opportunitiesContributors}          | ${notAuthorizedCode}
-    ${opportunitiesGroups}                | ${notAuthorizedCode}
     ${opportunitiesActorgroupsName}       | ${notAuthorizedCode}
     ${opportunitiesActorGroupsActorsName} | ${notAuthorizedCode}
     ${opportunitiesAspects}               | ${notAuthorizedCode}
     ${opportunitiesRelationsName}         | ${notAuthorizedCode}
-    ${projectsName}                       | ${notAuthorizedCode}
-    ${projectsTextId}                     | ${notAuthorizedCode}
-    ${projectsDescription}                | ${notAuthorizedCode}
-    ${projectsState}                      | ${notAuthorizedCode}
-    ${projectsTagset}                     | ${notAuthorizedCode}
-    ${projectsAspects}                    | ${notAuthorizedCode}
     ${templatesName}                      | ${notAuthorizedCode}
     ${templatesDescription}               | ${notAuthorizedCode}
   `(
@@ -266,11 +293,9 @@ describe('DDT community admin user - queries - authorized', () => {
         query: `${query}`,
         variables: null,
       };
-      const response = await graphqlRequestAuth(
-        requestParamsQueryData,
-        TestUser.GLOBAL_ADMIN
-      );
+      const response = await graphqlRequest(requestParamsQueryData);
       let responseData = JSON.stringify(response.body).replace('\\', '');
+      // 
 
       // Assert
       expect(response.status).toBe(200);
@@ -279,20 +304,15 @@ describe('DDT community admin user - queries - authorized', () => {
   );
 });
 
-// Failing scenarios BUG: https://app.zenhub.com/workspaces/cherrytwist-5ecb98b262ebd9f4aec4194c/issues/cherrytwist/server/578
-// ${opportunitiesProjectsId}          | ${notAuthorizedCode}
-// ${opportunitiesProjectsAspectsId}   | ${notAuthorizedCode}
-
-describe('DDT global admin user - Create mutations - authorized', () => {
+describe('DDT anonymous user - Create mutations - Not authorized', () => {
   // Arrange
   test.each`
     mutation                             | variables                             | expected
     ${createOrganisationMutation}        | ${createOrganisationVariables}        | ${notAuthorizedCode}
+    ${createGroupOnEcoverseMutation}     | ${createGroupOnEcoverseVariables}     | ${notAuthorizedCode}
+    ${createOrganisationMutation}        | ${createOrganisationVariables}        | ${notAuthorizedCode}
     ${createUserMutation}                | ${createUserVariables}                | ${notAuthorizedCode}
     ${createReferenceOnProfileMutation}  | ${createReferenceOnProfileVariable}   | ${notAuthorizedCode}
-    ${createReferenceOnContextMutation}  | ${createReferenceOnContextVariables}  | ${notAuthorizedCode}
-    ${createTagsetOnProfileMutation}     | ${createTagsetOnProfileVariables}     | ${notAuthorizedCode}
-    ${createGroupOnEcoverseMutation}     | ${createGroupOnEcoverseVariables}     | ${notAuthorizedCode}
     ${createTemplateMutation}            | ${createTemplateVariables}            | ${notAuthorizedCode}
     ${createChallengeMutation}           | ${createChallengeVariables}           | ${notAuthorizedCode}
     ${createGroupOnChallengeMutation}    | ${createGroupOnChallengeVariables}    | ${notAuthorizedCode}
@@ -304,6 +324,8 @@ describe('DDT global admin user - Create mutations - authorized', () => {
     ${createAspectOnOpportunityMutation} | ${createAspectOnOpportunityVariables} | ${notAuthorizedCode}
     ${createRelationMutation}            | ${createRelationVariables}            | ${notAuthorizedCode}
     ${createAspectOnProjectMutation}     | ${createAspectOnProjectVariables}     | ${notAuthorizedCode}
+    ${createReferenceOnContextMutation}  | ${createReferenceOnContextVariables}  | ${notAuthorizedCode}
+    ${createTagsetOnProfileMutation}     | ${createTagsetOnProfileVariables}     | ${notAuthorizedCode}
   `(
     "should expect: '$expected' for create mutation: '$mutation' and variables: '$variables'",
     async ({ mutation, variables, expected }) => {
@@ -313,26 +335,28 @@ describe('DDT global admin user - Create mutations - authorized', () => {
         query: `${mutation}`,
         variables: `${variables}`,
       };
-      const response = await graphqlRequestAuth(
-        requestParamsCreateMutations,
-        TestUser.GLOBAL_ADMIN
-      );
+      const response = await graphqlRequest(requestParamsCreateMutations);
       let responseData = JSON.stringify(response.body).replace('\\', '');
+      
 
       // Assert
       expect(response.status).toBe(200);
-      expect(responseData).not.toContain(expected);
+      expect(responseData).toContain(expected);
     }
   );
 });
 
-describe('DDT global admin user - Update mutations - authorized', () => {
+describe('DDT anonymous user - Update mutations - NOT authorized', () => {
   // Arrange
   test.each`
     mutation                                  | variables                                  | expected
     ${updateUserMutation}                     | ${updateUserVariables}                     | ${notAuthorizedCode}
     ${updateProfileMutation}                  | ${updateProfileVariables}                  | ${notAuthorizedCode}
     ${updateOrganisationMutation}             | ${updateOrganisationVariabls}              | ${notAuthorizedCode}
+    ${updateChallengeMutation}                | ${updateChallengeVariables}                | ${notAuthorizedCode}
+    ${updateOpportunityMutation}              | ${updateOpportunityVariables}              | ${notAuthorizedCode}
+    ${updateAspectMutation}                   | ${updateAspectVariable}                    | ${notAuthorizedCode}
+    ${updateActorMutation}                    | ${updateActorVariables}                    | ${notAuthorizedCode}
     ${addTagsOnTagsetMutation}                | ${addTagsOnTagsetVariables}                | ${notAuthorizedCode}
     ${replaceTagsOnTagsetMutation}            | ${replaceTagsOnTagsetVariables}            | ${notAuthorizedCode}
     ${addUserToChallengeMutation}             | ${addUserToChallengeVariables}             | ${notAuthorizedCode}
@@ -342,10 +366,6 @@ describe('DDT global admin user - Update mutations - authorized', () => {
     ${removeGroupFocalPointMutation}          | ${removeGroupFocalPointVariables}          | ${notAuthorizedCode}
     ${addChallengeLeadToOrganisationMutation} | ${addChallengeLeadToOrganisationVariables} | ${notAuthorizedCode}
     ${removeUserFromGroupMutation}            | ${removeUserFromGroupVariables}            | ${notAuthorizedCode}
-    ${updateChallengeMutation}                | ${updateChallengeVariables}                | ${notAuthorizedCode}
-    ${updateOpportunityMutation}              | ${updateOpportunityVariables}              | ${notAuthorizedCode}
-    ${updateAspectMutation}                   | ${updateAspectVariable}                    | ${notAuthorizedCode}
-    ${updateActorMutation}                    | ${updateActorVariables}                    | ${notAuthorizedCode}
   `(
     "should expect: '$expected' for update mutation: '$mutation' and variables: '$variables'",
     async ({ mutation, variables, expected }) => {
@@ -355,28 +375,26 @@ describe('DDT global admin user - Update mutations - authorized', () => {
         query: `${mutation}`,
         variables: `${variables}`,
       };
-      const response = await graphqlRequestAuth(
-        requestParamsUpdateMutations,
-        TestUser.GLOBAL_ADMIN
-      );
+      const response = await graphqlRequest(requestParamsUpdateMutations);
       let responseData = JSON.stringify(response.body).replace('\\', '');
+      
 
       // Assert
       expect(response.status).toBe(200);
-      expect(responseData).not.toContain(expected);
+      expect(responseData).toContain(expected);
     }
   );
 });
 
-describe('DDT global admin user - Remove mutations - authorized', () => {
+describe('DDT anonymous user - Remove mutations - NOT authorized', () => {
   // Arrange
   test.each`
     mutation                    | variables                    | expected
-    ${removeActorGroupMutation} | ${removeActorGroupVariables} | ${notAuthorizedCode}
-    ${removeActorMutation}      | ${removeActorVariables}      | ${notAuthorizedCode}
-    ${removeAspectMutation}     | ${removeAspectVariables}     | ${notAuthorizedCode}
-    ${removeChallengeMutation}  | ${removeChallengeVariables}  | ${notAuthorizedCode}
     ${removeUserMutation}       | ${removeUserVariables}       | ${notAuthorizedCode}
+    ${removeChallengeMutation}  | ${removeChallengeVariables}  | ${notAuthorizedCode}
+    ${removeAspectMutation}     | ${removeAspectVariables}     | ${notAuthorizedCode}
+    ${removeActorMutation}      | ${removeActorVariables}      | ${notAuthorizedCode}
+    ${removeActorGroupMutation} | ${removeActorGroupVariables} | ${notAuthorizedCode}
   `(
     "should expect: '$expected' for remove mutation: '$mutation' and variables: '$variables'",
     async ({ mutation, variables, expected }) => {
@@ -386,15 +404,13 @@ describe('DDT global admin user - Remove mutations - authorized', () => {
         query: `${mutation}`,
         variables: `${variables}`,
       };
-      const response = await graphqlRequestAuth(
-        requestParamsRemoveMutations,
-        TestUser.GLOBAL_ADMIN
-      );
+      const response = await graphqlRequest(requestParamsRemoveMutations);
       let responseData = JSON.stringify(response.body).replace('\\', '');
+      
 
       // Assert
       expect(response.status).toBe(200);
-      expect(responseData).not.toContain(expected);
+      expect(responseData).toContain(expected);
     }
   );
 });
