+++ test/ci-build/auth/queries.ts
