+++ test/ci-build/challenge/challenge.request.params.ts
@@ -1,5 +1,5 @@
-import { TestUser } from '../utils/token.helper';
-import { graphqlRequestAuth } from '../utils/graphql.request';
+import { TestUser } from '../../utils/token.helper';
+import { graphqlRequestAuth } from '../../utils/graphql.request';
 
 const uniqueId = (Date.now() + Math.random()).toString();
